pyproject.toml
[project]
name = "agentick-be"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "email-validator>=2.3.0",
    "fastapi>=0.135.3",
    "psycopg[binary]>=3.3.3",
    "pwdlib[argon2]>=0.3.0",
    "pyjwt>=2.10.1",
    "pydantic-settings>=2.13.1",
    "sqlalchemy>=2.0.49",
    "uvicorn[standard]>=0.44.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.3",
    "ruff>=0.15.9",
]
